cleanup for version, bump 0.9
