+++ bot/skills/__init__.py
@@ -4,6 +4,7 @@ from typing import List, Dict, Callable
 from telegram import Update
 from telegram.ext import CommandHandler, Updater, CallbackContext, run_async
 
+from mode import cleanup
 from skills.core import add_core
 from skills.mute import add_mute
 from skills.roll import add_roll
@@ -14,8 +15,9 @@ from skills.towel_mode import add_towel_mode
 from skills.tree import add_tree
 from skills.uwu import add_uwu
 
-__version__ = "0.8.9"
 from filters import admin_filter
+
+__version__ = "0.9"
 
 logger = logging.getLogger(__name__)
 
@@ -27,6 +29,8 @@ def _add_version(upd: Updater, version_handlers_group: int):
 
 
 @run_async
+@cleanup(seconds=2)
+# https://github.com/egregors/vldc-bot/issues/80
 def _version(update: Update, context: CallbackContext):
     """ Show current version of bot """
     logger.info(f"current ver.: {__version__}")
